Scrabble-Like Game using Binary Search
Working scrabble-like game that assigns two players 4 random letters and asks for them to both create the longest possible word from the available letters. The player's words will first be checked using the txt file. If both words are valid the player with the higher-scoring word, based on true Scrabble scoring, will win. If one player's word is not valid they will lose, and if both words are not valid or are equally scored the players will tie. Players also have the option to end the game at any time or exchange a letter at a certain index.